backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
piexif==1.1.3
geopy==2.4.1
python-dotenv==1.0.0
